requirements.txt
Package                      Version
---------------------------- -----------
absl-py                      2.1.0
astunparse                   1.6.3
attrs                        24.2.0
cachetools                   5.5.0
certifi                      2024.8.30
cffi                         1.17.1
charset-normalizer           3.4.0
contourpy                    1.1.1
cycler                       0.12.1
flatbuffers                  24.3.25
fonttools                    4.55.2
gast                         0.4.0
google-auth                  2.36.0
google-auth-oauthlib         1.0.0
google-pasta                 0.2.0
grpcio                       1.68.1
h5py                         3.11.0
idna                         3.10
importlib_metadata           8.5.0
importlib_resources          6.4.5
jax                          0.4.13
keras                        2.13.1
kiwisolver                   1.4.7
libclang                     18.1.1
Markdown                     3.7
MarkupSafe                   2.1.5
matplotlib                   3.7.5
mediapipe                    0.10.11
ml-dtypes                    0.2.0
numpy                        1.22.0
oauthlib                     3.2.2
opencv-contrib-python        4.10.0.84
opencv-python                4.10.0.84
opt_einsum                   3.4.0
packaging                    24.2
paho-mqtt                    1.0
pillow                       10.4.0
pip                          24.2
protobuf                     3.20.3
pyasn1                       0.6.1
pyasn1_modules               0.4.1
pycparser                    2.22
pyparsing                    3.1.4
python-dateutil              2.9.0.post0
requests                     2.32.3
requests-oauthlib            2.0.0
rsa                          4.9
scipy                        1.10.1
setuptools                   75.1.0
six                          1.17.0
sounddevice                  0.5.1
tensorboard                  2.13.0
tensorboard-data-server      0.7.2
tensorflow                   2.13.0
tensorflow-estimator         2.13.0
tensorflow-intel             2.13.0
tensorflow-io-gcs-filesystem 0.31.0
termcolor                    2.4.0
typing_extensions            4.5.0
urllib3                      2.2.3
Werkzeug                     3.0.6
wheel                        0.44.0
wrapt                        1.17.0
zipp                         3.20.2
